Move Color extension to a dedicated file
