+++ Sources/UIComponents/Extensions/Color+hex.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+@available(iOS 13, macOS 10.15, *)
+internal extension Color {
+    init(hex: UInt, alpha: Double = 1) {
+        self.init(
+            .sRGB,
+            red: Double((hex >> 16) & 0xFF) / 255,
+            green: Double((hex >> 08) & 0xFF) / 255,
+            blue: Double((hex >> 00) & 0xFF) / 255
+        )
+    }
+}
