+++ Sources/UIComponents/PokemonType.swift
@@ -94,18 +94,6 @@ public enum PokemonType: String, CaseIterable {
     }
 }
 
-@available(iOS 13, macOS 10.15, *)
-internal extension Color {
-    init(hex: UInt, alpha: Double = 1) {
-        self.init(
-            .sRGB,
-            red: Double((hex >> 16) & 0xFF) / 255,
-            green: Double((hex >> 08) & 0xFF) / 255,
-            blue: Double((hex >> 00) & 0xFF) / 255
-        )
-    }
-}
-
 #if DEBUG
 @available(iOS 13, macOS 11.0, *)
 struct PokemonTypeColor_Previews: PreviewProvider {
